Overhaul app layout
This commit includes lots of changes, mostly focused on overhauling
the layout system of the site. Previously it was a mishmash of CSS grid
rules and exceptions, with different states for a sidebar page, a full-
width page, or both/neither, and it got to be unmanageable.

This update simplifies things dramatically. Instead of designating full-
width pages, constrained, centered pages simply add utility classes to
themselves, and fullwidth pages adhere to one basic media query and
layout.

This commit also updates the homepage to be more interesting at wide
screen widths, and contains various other minor updates, such as moving
the location of the page head to always be beneath the logo.